requirements.txt
aiohttp==3.12.14
aiohttp_socks==0.10.1
beautifulsoup4==4.13.4
elasticsearch==9.0.2
nltk==3.9.1
numpy==2.3.1
Pillow==11.3.0
redis==6.2.0
scikit_learn==1.7.1
selenium==4.34.2
torch==2.7.1
torchvision==0.22.1
transformers==4.53.2
